change lock

+++ build/lib/drf_advanced_token_manager/views.py
@@ -5,8 +5,6 @@ from django.shortcuts import reverse
 from django.views import View
 from django.contrib.auth.mixins import LoginRequiredMixin
 
-from django_app_permissions.views import APIAppAuthView
-from django_app_permissions.views import AppAuthView
 from django.contrib import messages
 from django.conf import settings
 from django.contrib.auth.models import User, Group
@@ -16,16 +14,24 @@ from django.conf import settings
 from django.http import Http404, HttpResponse, HttpResponseForbidden
 from django.core.exceptions import ObjectDoesNotExist
 import uuid
+from drf_advanced_token_manager.models import UserUIKeyLock
 
 class TokenView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         try:
             token = Token.objects.get(user=request.user)
         except ObjectDoesNotExist as odne:
-            Token.objects.create(user=request.user, key=str(uuid.uuid4()))
+            token = Token.objects.create(user=request.user, key=str(uuid.uuid4()))
+
+        home_url = reverse(settings.SITE_HOME_URL_NAME) if hasattr(settings,"SITE_HOME_URL_NAME") else "/"
+
+        if (hasattr(settings,"PREVENT_TOKEN_UI_CHANGE") and settings.PREVENT_TOKEN_UI_CHANGE) or UserUIKeyLock.objects.filter(user=request.user).exists():
+            change_lock = True
+        else:
+            change_lock = False
 
         return render(request, "drf_token_management_view_token.html",
-         {"token":token,"site_home":reverse(settings.SITE_HOME_URL_NAME)})
+         {"token":token,"site_home":home_url, "change_lock":change_lock})
         
 
     def post(self, request, *args, **kwargs):
@@ -35,13 +41,19 @@ class TokenView(LoginRequiredMixin, View):
 class ChangeTokenView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         # are you sure page
+        if UserUIKeyLock.objects.filter(user=request.user).exists():
+            print("locked")
+            return HttpResponseForbidden("You are not allowed to change your token :P")
         if hasattr(settings,"PREVENT_TOKEN_UI_CHANGE") and settings.PREVENT_TOKEN_UI_CHANGE:
-            return render(request, "drf_token_management_change_token.html", 
-              {"site_home":reverse(settings.SITE_HOME_URL_NAME)})
-        else:
-            return HttpResponseForbidden()
+            print("token change locked")
+            return HttpResponseForbidden("You are not allowed to change your token :P")
+        home_url = reverse(settings.SITE_HOME_URL_NAME) if hasattr(settings,"SITE_HOME_URL_NAME") else "/"
+        return render(request, "drf_token_management_change_token.html", 
+        {"site_home":home_url})
 
     def post(self, request, *args, **kwargs):
+        if UserUIKeyLock.objects.filter(user=request.user).exists():
+            return HttpResponseForbidden("You are not allowed to change your token :P")
         token = Token.objects.get(user=request.user)
         token.delete()
         token = Token.objects.create(user=request.user, key=str(uuid.uuid4()))
